Simplify estimate-drawing code, unifying Zen and Kata and on/off with mode-cycling.

+++ src/main/java/featurecat/lizzie/gui/MainFrame.java
@@ -319,7 +319,7 @@ public abstract class MainFrame extends JFrame {
 
   public abstract void noEstimateByZen();
 
-  public abstract void drawEstimateRectZen(ArrayList<Integer> esitmateArray);
+  public abstract void drawEstimateRectZen(ArrayList<Double> esitmateArray);
 
   public void saveImage() {};
 }
